extract helpers into helpers_test
